+++ src/main/java/com/asu/seatr/api/CourseAPI_1.java
@@ -22,70 +22,74 @@ import com.asu.seatr.auth.AuthenticationService;
 import com.asu.seatr.exceptions.CourseException;
 import com.asu.seatr.exceptions.UserException;
 import com.asu.seatr.handlers.CourseAnalyzerHandler;
-import com.asu.seatr.handlers.CourseHandler;
 import com.asu.seatr.handlers.UserCourseHandler;
-import com.asu.seatr.handlers.UserHandler;
-import com.asu.seatr.models.Course;
-import com.asu.seatr.models.User;
-import com.asu.seatr.models.UserCourse;
 import com.asu.seatr.models.analyzers.course.C_A1;
 import com.asu.seatr.rest.models.CAReader1;
 import com.asu.seatr.utils.MyMessage;
 import com.asu.seatr.utils.MyResponse;
 import com.asu.seatr.utils.MyStatus;
+import com.asu.seatr.utils.Utilities;
 
 @Path("/analyzer/1/courses")
 public class CourseAPI_1 {
 	public static final String AUTHENTICATION_HEADER = "Authorization";
 
 	static Logger logger = Logger.getLogger(CourseAPI_1.class);
+
 	@GET
 	@Produces(MediaType.APPLICATION_JSON)
 	public CAReader1 getCourse(@QueryParam("external_course_id") String external_course_id){
-		C_A1 ca1 = null;
+
 		try {
-			ca1 = (C_A1)CourseAnalyzerHandler.readByExtId(C_A1.class, external_course_id).get(0);
-		} catch (CourseException e) {
-			Response rb = Response.status(Status.NOT_FOUND)
-					.entity(MyResponse.build(e.getMyStatus(), e.getMyMessage())).build();			
-			throw new WebApplicationException(rb);
-		} catch(Exception e){
-			logger.error(e.getStackTrace());
-			System.out.println(e.getMessage());
-			Response rb = Response.status(Status.BAD_REQUEST)
-					.entity(MyResponse.build(MyStatus.ERROR, MyMessage.BAD_REQUEST)).build();
-			throw new WebApplicationException(rb);
-		}
-		
+			if(!Utilities.checkExists(external_course_id)) {
+				throw new CourseException(MyStatus.ERROR, MyMessage.COURSE_ID_MISSING);
+			}
+
+			C_A1 ca1 = (C_A1)CourseAnalyzerHandler.readByExtId(C_A1.class, external_course_id).get(0);
 			CAReader1 reader=new CAReader1();
 			reader.setDescription(ca1.getCourseDesc());
 			reader.setExternal_course_id(ca1.getCourseExtId());
 			reader.setTeaching_unit(ca1.getTeaching_unit());
 			reader.setThreshold(ca1.getThreshold());
 			return reader;
-		
+		} catch (CourseException e) {
+			Response rb = Response.status(Status.NOT_FOUND)
+					.entity(MyResponse.build(e.getMyStatus(), e.getMyMessage())).build();			
+			throw new WebApplicationException(rb);
+		} catch(Exception e){
+			logger.error("Exception while getting course -analyzer 1", e);			
+			Response rb = Response.status(Status.BAD_REQUEST)
+					.entity(MyResponse.build(MyStatus.ERROR, MyMessage.BAD_REQUEST)).build();
+			throw new WebApplicationException(rb);
+		}
+
+
+
 	}
-		
+
 	@POST
 	@Consumes(MediaType.APPLICATION_JSON)
 	@Produces(MediaType.APPLICATION_JSON)
 	public Response createCourse(CAReader1 reader, @HeaderParam(AUTHENTICATION_HEADER) String authHeader){
-		
-		C_A1 ca1 = new C_A1();
-		StringTokenizer tokenizer = AuthenticationService.getUsernameAndPassword(authHeader);
-		String username = tokenizer.nextToken();
-		
 		try{
+			if(!Utilities.checkExists(reader.getExternal_course_id())) {
+				throw new CourseException(MyStatus.ERROR, MyMessage.COURSE_ID_MISSING);
+			}		
+
+			C_A1 ca1 = new C_A1();
+			StringTokenizer tokenizer = AuthenticationService.getUsernameAndPassword(authHeader);
+			String username = tokenizer.nextToken();		
+
 			ca1.createCourse(reader.getExternal_course_id(), reader.getDescription());
 			ca1.setTeaching_unit(reader.getTeaching_unit());
 			ca1.setThreshold(reader.getThreshold());
 			CourseAnalyzerHandler.save(ca1);
-			
+
 			UserCourseHandler.save(username, reader.getExternal_course_id());
-			
+
 			return Response.status(Status.CREATED)
 					.entity(MyResponse.build(MyStatus.SUCCESS, MyMessage.COURSE_CREATED)).build();
-		
+
 		} catch (CourseException  e) {
 			Response rb = Response.status(Status.OK)
 					.entity(MyResponse.build(e.getMyStatus(), e.getMyMessage())).build();			
@@ -96,25 +100,33 @@ public class CourseAPI_1 {
 			throw new WebApplicationException(rb);
 		}
 		catch(Exception e){
-			logger.error(e.getStackTrace());
+			logger.error("Exception while creating course", e);
 			Response rb = Response.status(Status.BAD_REQUEST)
 					.entity(MyResponse.build(MyStatus.ERROR, MyMessage.BAD_REQUEST)).build();
 			throw new WebApplicationException(rb);
-		}
-								
-			
+		}								
+
 	}
-		
+
 	@PUT
 	@Consumes(MediaType.APPLICATION_JSON)
 	@Produces(MediaType.APPLICATION_JSON)
 	public Response updateCourse(CAReader1 reader){
 		try{
-			
+			if(!Utilities.checkExists(reader.getExternal_course_id())) {
+				throw new CourseException(MyStatus.ERROR, MyMessage.COURSE_ID_MISSING);
+			}
+
+
 			C_A1 ca1 = (C_A1) CourseAnalyzerHandler.readByExtId(C_A1.class, reader.getExternal_course_id()).get(0);
-			
-			ca1.setTeaching_unit(reader.getTeaching_unit());
-			ca1.setThreshold(reader.getThreshold());
+
+			if(Utilities.checkExists(reader.getTeaching_unit())) {
+				ca1.setTeaching_unit(reader.getTeaching_unit());
+			}	
+			if(Utilities.checkExists(reader.getThreshold())) {
+				ca1.setThreshold(reader.getThreshold());
+			}
+
 			CourseAnalyzerHandler.update(ca1);
 			return Response.status(Status.OK)
 					.entity(MyResponse.build(MyStatus.SUCCESS, MyMessage.COURSE_UPDATED))
@@ -124,23 +136,24 @@ public class CourseAPI_1 {
 					.entity(MyResponse.build(e.getMyStatus(), e.getMyMessage())).build();			
 			throw new WebApplicationException(rb);
 		} catch(Exception e){
-			logger.error(e.getStackTrace());
+			logger.error("Exception while updating course", e);
 			Response rb = Response.status(Status.BAD_REQUEST)
 					.entity(MyResponse.build(MyStatus.ERROR, MyMessage.BAD_REQUEST)).build();
 			throw new WebApplicationException(rb);
 		}
 	}
-	
-			
+
+
 	@DELETE
 	@Consumes(MediaType.APPLICATION_JSON)
 	@Produces(MediaType.APPLICATION_JSON)
 	public Response deleteCourseAnalyzer1(@QueryParam("external_course_id") String external_course_id) {
-		
-		C_A1 c_a1;
-		// delete all course analyzers and then delete the course record
+
 		try {
-			c_a1 = (C_A1)CourseAnalyzerHandler.readByExtId(C_A1.class, external_course_id).get(0);
+			if(!Utilities.checkExists(external_course_id)) {
+				throw new CourseException(MyStatus.ERROR, MyMessage.COURSE_ID_MISSING);
+			}
+			C_A1 c_a1 = (C_A1)CourseAnalyzerHandler.readByExtId(C_A1.class, external_course_id).get(0);
 			CourseAnalyzerHandler.delete(c_a1);
 			return Response.status(Status.OK)
 					.entity(MyResponse.build(MyStatus.SUCCESS, MyMessage.COURSE_ANALYZER_DELETED)).build();
@@ -149,13 +162,13 @@ public class CourseAPI_1 {
 					.entity(MyResponse.build(e.getMyStatus(), e.getMyMessage())).build();			
 			throw new WebApplicationException(rb);
 		} catch(Exception e){
-			logger.error(e.getStackTrace());
+			logger.error("Exception while deleting course analyzer", e);
 			Response rb = Response.status(Status.BAD_REQUEST)
 					.entity(MyResponse.build(MyStatus.ERROR, MyMessage.BAD_REQUEST)).build();
 			throw new WebApplicationException(rb);
 		}
 	}
-	
-	
-		
+
+
+
 }
