+++ src/main/java/com/asu/seatr/api/CourseAPI_3.java
@@ -22,75 +22,75 @@ import com.asu.seatr.auth.AuthenticationService;
 import com.asu.seatr.exceptions.CourseException;
 import com.asu.seatr.exceptions.UserException;
 import com.asu.seatr.handlers.CourseAnalyzerHandler;
-import com.asu.seatr.handlers.CourseHandler;
 import com.asu.seatr.handlers.UserCourseHandler;
-import com.asu.seatr.handlers.UserHandler;
-import com.asu.seatr.models.Course;
-import com.asu.seatr.models.User;
-import com.asu.seatr.models.UserCourse;
-import com.asu.seatr.models.analyzers.course.C_A1;
 import com.asu.seatr.models.analyzers.course.C_A3;
-import com.asu.seatr.rest.models.CAReader1;
 import com.asu.seatr.rest.models.CAReader3;
 import com.asu.seatr.utils.MyMessage;
 import com.asu.seatr.utils.MyResponse;
 import com.asu.seatr.utils.MyStatus;
+import com.asu.seatr.utils.Utilities;
 
 @Path("/analyzer/3/courses")
 public class CourseAPI_3 {
 	public static final String AUTHENTICATION_HEADER = "Authorization";
 
-	static Logger logger = Logger.getLogger(CourseAPI_1.class);
+	static Logger logger = Logger.getLogger(CourseAPI_3.class);
 	@GET
 	@Produces(MediaType.APPLICATION_JSON)
-	public CAReader3 getCourse(@QueryParam("external_course_id") String external_course_id){
-		C_A3 ca3 = null;
+	public CAReader3 getCourse(@QueryParam("external_course_id") String external_course_id){		
 		try {
-			ca3 = (C_A3)CourseAnalyzerHandler.readByExtId(C_A3.class, external_course_id).get(0);
-		} catch (CourseException e) {
-			Response rb = Response.status(Status.NOT_FOUND)
-					.entity(MyResponse.build(e.getMyStatus(), e.getMyMessage())).build();			
-			throw new WebApplicationException(rb);
-		} catch(Exception e){
-			logger.error(e.getStackTrace());
-			System.out.println(e.getMessage());
-			Response rb = Response.status(Status.BAD_REQUEST)
-					.entity(MyResponse.build(MyStatus.ERROR, MyMessage.BAD_REQUEST)).build();
-			throw new WebApplicationException(rb);
-		}
-		
+			if(!Utilities.checkExists(external_course_id)) {
+				throw new CourseException(MyStatus.ERROR, MyMessage.COURSE_ID_MISSING);
+			}
+			C_A3 ca3  = (C_A3)CourseAnalyzerHandler.readByExtId(C_A3.class, external_course_id).get(0);
 			CAReader3 reader=new CAReader3();
 			reader.setDescription(ca3.getCourseDesc());
 			reader.setExternal_course_id(ca3.getCourseExtId());
 			reader.setD_current_unit_no(ca3.getD_current_unit_no());
 			reader.setD_max_n(ca3.getD_max_n());
 			reader.setS_units(ca3.getS_units());
-			
+
 			return reader;
-		
+		} catch (CourseException e) {
+			Response rb = Response.status(Status.NOT_FOUND)
+					.entity(MyResponse.build(e.getMyStatus(), e.getMyMessage())).build();			
+			throw new WebApplicationException(rb);
+		} catch(Exception e){
+			logger.error("Exception while getting course - analyzer 3", e);			
+			Response rb = Response.status(Status.BAD_REQUEST)
+					.entity(MyResponse.build(MyStatus.ERROR, MyMessage.BAD_REQUEST)).build();
+			throw new WebApplicationException(rb);
+		}
+
+
+
 	}
-		
+
 	@POST
 	@Consumes(MediaType.APPLICATION_JSON)
 	@Produces(MediaType.APPLICATION_JSON)
 	public Response createCourse(CAReader3 reader, @HeaderParam(AUTHENTICATION_HEADER) String authHeader){
-		
-		C_A3 ca3 = new C_A3();
-		StringTokenizer tokenizer = AuthenticationService.getUsernameAndPassword(authHeader);
-		String username = tokenizer.nextToken();
-		
+
 		try{
+			if(!Utilities.checkExists(reader.getExternal_course_id())) {
+				throw new CourseException(MyStatus.ERROR, MyMessage.COURSE_ID_MISSING);
+			}
+
+			C_A3 ca3 = new C_A3();
+			StringTokenizer tokenizer = AuthenticationService.getUsernameAndPassword(authHeader);
+			String username = tokenizer.nextToken();
+
 			ca3.createCourse(reader.getExternal_course_id(), reader.getDescription());
 			ca3.setD_current_unit_no(reader.getD_current_unit_no());
 			ca3.setD_max_n(reader.getD_max_n());
 			ca3.setS_units(reader.getS_units());
 			CourseAnalyzerHandler.save(ca3);
-			
+
 			UserCourseHandler.save(username, reader.getExternal_course_id());
-			
+
 			return Response.status(Status.CREATED)
 					.entity(MyResponse.build(MyStatus.SUCCESS, MyMessage.COURSE_CREATED)).build();
-		
+
 		} catch (CourseException  e) {
 			Response rb = Response.status(Status.OK)
 					.entity(MyResponse.build(e.getMyStatus(), e.getMyMessage())).build();			
@@ -101,27 +101,37 @@ public class CourseAPI_3 {
 			throw new WebApplicationException(rb);
 		}
 		catch(Exception e){
-			logger.error(e.getStackTrace());
+			logger.error("Exception while creating course", e);
 			Response rb = Response.status(Status.BAD_REQUEST)
 					.entity(MyResponse.build(MyStatus.ERROR, MyMessage.BAD_REQUEST)).build();
 			throw new WebApplicationException(rb);
 		}
-								
-			
+
+
 	}
-		
+
 	@PUT
 	@Consumes(MediaType.APPLICATION_JSON)
 	@Produces(MediaType.APPLICATION_JSON)
 	public Response updateCourse(CAReader3 reader){
 		try{
-			
+			if(!Utilities.checkExists(reader.getExternal_course_id())) {
+				throw new CourseException(MyStatus.ERROR, MyMessage.COURSE_ID_MISSING);
+			}
+
+
 			C_A3 ca3 = (C_A3) CourseAnalyzerHandler.readByExtId(C_A3.class, reader.getExternal_course_id()).get(0);
-			
-			ca3.setD_current_unit_no(reader.getD_current_unit_no());
-			ca3.setD_max_n(reader.getD_max_n());
-			ca3.setS_units(reader.getS_units());
-			
+
+			if(Utilities.checkExists(reader.getD_current_unit_no())) {
+				ca3.setD_current_unit_no(reader.getD_current_unit_no());
+			}
+			if(Utilities.checkExists(reader.getD_max_n())) {
+				ca3.setD_max_n(reader.getD_max_n());
+			}
+			if(Utilities.checkExists(reader.getS_units())) {
+				ca3.setS_units(reader.getS_units());
+			}
+
 			CourseAnalyzerHandler.update(ca3);
 			return Response.status(Status.OK)
 					.entity(MyResponse.build(MyStatus.SUCCESS, MyMessage.COURSE_UPDATED))
@@ -132,24 +142,24 @@ public class CourseAPI_3 {
 			throw new WebApplicationException(rb);
 		} 
 		catch(Exception e){
-			System.out.println(e.getStackTrace());
-			logger.error(e.getStackTrace());
+			logger.error("Exception while updating course", e);
 			Response rb = Response.status(Status.BAD_REQUEST)
 					.entity(MyResponse.build(MyStatus.ERROR, MyMessage.BAD_REQUEST)).build();
 			throw new WebApplicationException(rb);
 		}
 	}
-	
-			
+
+
 	@DELETE
 	@Consumes(MediaType.APPLICATION_JSON)
 	@Produces(MediaType.APPLICATION_JSON)
 	public Response deleteCourseAnalyzer1(@QueryParam("external_course_id") String external_course_id) {
-		
-		C_A3 c_a3;
-		// delete all course analyzers and then delete the course record
+
 		try {
-			c_a3 = (C_A3)CourseAnalyzerHandler.readByExtId(C_A3.class, external_course_id).get(0);
+			if(!Utilities.checkExists(external_course_id)) {
+				throw new CourseException(MyStatus.ERROR, MyMessage.COURSE_ID_MISSING);
+			}
+			C_A3 c_a3 = (C_A3)CourseAnalyzerHandler.readByExtId(C_A3.class, external_course_id).get(0);
 			CourseAnalyzerHandler.delete(c_a3);
 			return Response.status(Status.OK)
 					.entity(MyResponse.build(MyStatus.SUCCESS, MyMessage.COURSE_ANALYZER_DELETED)).build();
@@ -158,13 +168,13 @@ public class CourseAPI_3 {
 					.entity(MyResponse.build(e.getMyStatus(), e.getMyMessage())).build();			
 			throw new WebApplicationException(rb);
 		} catch(Exception e){
-			logger.error(e.getStackTrace());
+			logger.error("Exception while deleting course analyzer", e);
 			Response rb = Response.status(Status.BAD_REQUEST)
 					.entity(MyResponse.build(MyStatus.ERROR, MyMessage.BAD_REQUEST)).build();
 			throw new WebApplicationException(rb);
 		}
 	}
-	
-	
-		
+
+
+
 }
